refactor(protocol): move context/tools → context/outlets (PR-1 rename step 1)
