+++ crates/scp-protocol/src/context/outlets/interface.rs
