+++ crates/scp-protocol/src/context/outlets/registry.rs
