+++ crates/scp-protocol/src/context/outlets/schema.rs
