+++ crates/scp-protocol/src/context/outlets/summary.rs
